backend/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangGraph et LangChain
langgraph==0.2.39
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.2.4

# Base de données (optionnel)
sqlalchemy==2.0.23
alembic==1.13.0

# CORS et middleware
python-cors==1.7.0

# Utilitaires
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
